conf: Turn def->maxvcpus into size_t
Later on this will also be used to track size of the vcpu data array.
Use size_t so that we can utilize the memory allocation helpers.

+++ src/conf/domain_conf.h
@@ -2219,7 +2219,7 @@ struct _virDomainDef {
     virDomainMemtune mem;
 
     unsigned int vcpus;
-    unsigned int maxvcpus;
+    size_t maxvcpus;
     int placement_mode;
     virBitmapPtr cpumask;
